requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
dateparser==1.1.0
environs==9.5.0
fonttools==4.30.0
frozenlist==1.3.0
idna==3.3
kiwisolver==1.3.2
marshmallow==3.15.0
matplotlib==3.5.1
multidict==6.0.2
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pyparsing==3.0.7
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.27.1
six==1.16.0
TA-Lib==0.4.24
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.8
websockets==9.1
yarl==1.7.2
